--- a/sonar-project.properties
+++ b/sonar-project.properties
@@ -1,0 +1,17 @@
+# SonarCloud Configuration for Frontend Service
+
+sonar.organization=rpgoldberg
+sonar.projectKey=rpgoldberg_figure-collector-frontend
+sonar.projectName=Figure Collector Frontend
+
+sonar.sources=src
+sonar.exclusions=**/node_modules/**,**/build/**,**/*.test.tsx,**/*.test.ts,**/*.spec.tsx,**/setupTests.ts,**/__mocks__/**,**/test-utils.tsx,**/pages/Login.tsx,**/pages/Profile.tsx,**/pages/Register.tsx,**/pages/Statistics.tsx
+sonar.tests=src
+sonar.test.inclusions=**/*.test.tsx,**/*.test.ts,**/*.spec.tsx,**/*.spec.ts
+sonar.javascript.lcov.reportPaths=coverage/lcov.info
+sonar.typescript.lcov.reportPaths=coverage/lcov.info
+
+sonar.language=ts,tsx,js,jsx
+sonar.qualitygate.wait=true
+sonar.pullrequest.provider=github
+sonar.pullrequest.github.summary_comment=true