--- a/sonar-project.properties
+++ b/sonar-project.properties
@@ -5,9 +5,10 @@ sonar.projectKey=rpgoldberg_figure-collector-frontend
 sonar.projectName=Figure Collector Frontend
 
 sonar.sources=src
-sonar.exclusions=**/node_modules/**,**/build/**,**/*.test.tsx,**/*.test.ts,**/*.spec.tsx,**/setupTests.ts,**/__mocks__/**,**/test-utils.tsx,**/pages/Login.tsx,**/pages/Profile.tsx,**/pages/Register.tsx,**/pages/Statistics.tsx
+sonar.exclusions=**/node_modules/**,**/build/**,**/*.test.tsx,**/*.test.ts,**/*.spec.tsx,**/setupTests.ts,**/__mocks__/**,**/test-utils/**,**/test-utils.tsx,**/pages/Login.tsx,**/pages/Profile.tsx,**/pages/Register.tsx,**/pages/Statistics.tsx
 sonar.tests=src
 sonar.test.inclusions=**/*.test.tsx,**/*.test.ts,**/*.spec.tsx,**/*.spec.ts
+sonar.coverage.exclusions=**/test-utils/**,**/test-utils.tsx,**/*.test.tsx,**/*.test.ts,**/setupTests.ts,**/__mocks__/**,**/index.tsx,**/reportWebVitals.ts
 sonar.javascript.lcov.reportPaths=coverage/lcov.info
 sonar.typescript.lcov.reportPaths=coverage/lcov.info