--- a/src/__tests__/basic_math.test.ts
+++ b/src/__tests__/basic_math.test.ts
@@ -1,0 +1,5 @@
+describe('Basic Math', () => {
+  test('2 + 2 should equal 4', () => {
+    expect(2 + 2).toBe(4);
+  });
+});