--- a/src/pages/FigureList.tsx
+++ b/src/pages/FigureList.tsx
@@ -41,7 +41,7 @@ const FigureList: React.FC = () => {
         });
       },
     }
-  );
+  ) || { data: null, isLoading: false, error: null };
   
   const handlePageChange = (newPage: number) => {
     setPage(newPage);