… .claude/agents/documentation-manager.md
… .claude/agents/frontend-api-integrator.md
… .claude/agents/frontend-qa-automation-lead.md
… .claude/agents/frontend-state-manager.md
… .claude/agents/frontend-test-specialist.md
… .claude/agents/frontend-ui-composer.md
… .claude/commands/primer.md
… .dockerignore
… .github/workflows/build.yml
… .github/workflows/sonarcloud.yml
… .vscode/settings.json
… CLAUDE.md
… README.md
… TEST_COVERAGE_REPORT.md
… package.json
… src/__tests__/App.test.tsx
… src/__tests__/accessibility.test.tsx
… src/__tests__/e2e-workflows.test.tsx
… src/__tests__/integration/authentication.integration.test.tsx
… src/__tests__/integration/figure-operations.integration.test.tsx
… src/__tests__/integration/service-registration.integration.test.tsx
… src/__tests__/test-runner.js
… src/api/__tests__/index.enhanced.test.ts
… src/api/__tests__/index.test.ts
… src/components/FigureForm.tsx
… src/components/__tests__/EmptyState.test.tsx
… src/components/__tests__/FigureCard.test.tsx
… src/components/__tests__/FigureForm.comprehensive.test.tsx
… src/components/__tests__/FigureForm.conditions.test.tsx
… src/components/__tests__/FigureForm.enhanced.test.tsx